TPASPKT-1468 Map DataCite v4 namespace to version 4.3 XML schema in METS schema import
The namespace and schemaLocation for DataCite v4 is defined in the METS
XML schema. This schemaLocation is used when evaluating the namespace
for DataCite v4 tags when they are part of the METS document. As Metax
v3 currently outputs DataCite 4.3, we need to use this version for the
schemaLocation for DataCite in the METS metadata.